.history/requirements_20240715232904.txt
beautifulsoup4
pandas
pyaudio
pyttsx3
pynput
pywhatkit
requests
speechrecognition
wikipedia
streamlit 
google-generativeai 
python-dotenv